strands-weather-agent/weather_agent/requirements.txt
# Core Strands packages for AI agents
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# MCP (Model Context Protocol) support
mcp>=0.1.0

# AWS SDK for Bedrock integration (Strands uses this internally)
boto3>=1.34.0

# FastAPI for simple API server
fastapi>=0.115.0
uvicorn>=0.32.0

# HTTP client for MCP Streamable HTTP transport
httpx>=0.27.0

# Environment configuration
python-dotenv>=1.0.0

# Data validation (used by FastAPI)
pydantic>=2.5.0

# FastMCP - keep for existing MCP servers
fastmcp>=0.1.7

# Basic utilities
colorama>=0.4.6  # For colored console output
pyyaml>=6.0.1    # For configuration files if needed

# Observability and monitoring
langfuse>=3.22.0  # LLM observability platform with OTEL support